added min and max height to regions

+++ src/main/java/dev/fanger/mapgen/util/SeedGen.java
@@ -6,7 +6,7 @@ public class SeedGen {
         return (short) (Math.random() * 32767);
     }
 
-    public static double randomNumber(double x, double y, short z, int max) {
+    public static double randomNumber(double x, double y, short z, double max) {
 //        return ((Math.abs(x) + Math.abs(y)) * z) % max;
         return Math.random() * max;
     }
